Garments Shop Application: POS in C#.
A stand alone POS system developed in C#. Receptionist can generate a unique BAR-CODE for a specific item/product using the application. Receptionist can save and print the generated BAR-CODE and dispatch it on any product/item. Moreover, The Receptionist can perform the CRUD operations. The back is supported with LINQ for manipulation of the Database. The receptionist can calculate the accumulative bill of the item purchased by a potential buyer and print the bill just by clicking one simple button.